Ultimos cambios a los patrones
