+++ src/main/java/org/example/comprtamiento/Strategy.java
@@ -8,6 +8,17 @@ public class Strategy {
     /**
      * Permite definir un proposito pero dependiendo del contexto implementarlo.
      * */
+    private CompressionContext contexto;
+
+    public Strategy() {
+        contexto = new CompressionContext();
+    }
+
+    public void ejecutar(List<File> files) {
+        CompressionStrategy strategy = new ZipCompressionStrategy();
+        contexto.setCompressionStrategy(strategy);
+        contexto.createArchive(files);
+    }
 }
 
 
@@ -18,12 +29,14 @@ interface CompressionStrategy {
 class ZipCompressionStrategy implements CompressionStrategy {
     public void compressFiles(List<File> files) {
         // implementación del algoritmo de compresión ZIP
+        System.out.println("Comprimiendo los archivos a .zip");
     }
 }
 
 class RarCompressionStrategy implements CompressionStrategy {
     public void compressFiles(List<File> files) {
         // implementación del algoritmo de compresión RAR
+        System.out.println("Comprimiendo los archivos a .rar");
     }
 }
